+++ src/rules/UserRules.php
@@ -42,17 +42,6 @@ class UserRules
         $rules[] =
         [
             ['password', 'newPassword'],
-            'string',
-            'max' => $settings->maxLength,
-            'tooLong' => Craft::t(
-                'password-policy',
-                Craft::t('password-policy','Password can maximum contain {max} characters.'),
-                ['max' => $settings->maxLength]
-            ),
-        ];
-        $rules[] =
-        [
-            ['password', 'newPassword'],
             'match',
             'pattern' => PasswordPolicy::$plugin->passwords->generatePattern(),
             'message' => Craft::t(
@@ -60,6 +49,20 @@ class UserRules
                 'Your password must contain at least one of each of the following: '
             ) . PasswordPolicy::$plugin->passwords->generateMessage(),
         ];
+
+        if ($settings->maxLength > $settings->minLength) {
+            $rules[] =
+                [
+                    ['password', 'newPassword'],
+                    'string',
+                    'max' => $settings->maxLength,
+                    'tooLong' => Craft::t(
+                        'password-policy',
+                        Craft::t('password-policy','Password can maximum contain {max} characters.'),
+                        ['max' => $settings->maxLength]
+                    ),
+                ];
+        }
 
         if ($settings->pwned) {
             $rules[] = [['password', 'newPassword'], PwnedValidator::class];